feat: interval rework
