+++ src/interval/mod.rs
@@ -1,5 +1,3 @@
 pub mod base;
-pub mod basis;
 pub mod bound;
-pub mod calendar;
-pub mod naive;
+pub mod interval;
